requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-docx>=1.1.2
pydantic>=2.8.0
python-multipart
pdfplumber
python-dotenv
agno>=1.7.0
